Converts ReactCamera to react effects for redux
