+++ TTKDTrainingApp/App.js
@@ -6,14 +6,14 @@ import {Provider} from 'react-redux';
 
 import store from './src/redux/store';
 import HomeScreen from './src/components/HomeScreen';
-import ReactCameraConnector from './src/components/ReactCamera';
+import ReactCamera from './src/components/ReactCamera';
 import VideoWithControlsConnector from './src/components/VideoWithControls';
 import Content from './src/components/Content';
 
 const AppNavigator = createStackNavigator(
   {
     HomeScreen: HomeScreen,
-    ReactCamera: ReactCameraConnector,
+    ReactCamera: ReactCamera,
     VideoWithControls: VideoWithControlsConnector,
     Content: Content,
   },
