--- TTKDTrainingApp/src/components/ReactCamera/index.js
@@ -1,3 +1,0 @@
-import {ReactCameraConnector} from './ReactCameraConnector';
-
-export default ReactCameraConnector;
